requirements.txt
Jinja2>=3.1.0
pytest>=8.0.0
flatlib
google-genai
python-dotenv
PyYAML>=6.0
